requirements.txt
asgiref
Django
django-environ
django-filter
djangorestframework
sqlparse
tzdata
